Restructure utilities
 - Rename 'show' package to 'sonic-cli'
 - Create 'sonic-eeprom' Python package to house EEPROM-releated modules in an organized fashion
 - Create 'sonic-sfp' Python package to house SFP-related modules
 - All scripts now reference the above packages
 - sfputilbase.py now correctly parses 2- and 3-column port_config.ini/portmap.ini files
