+++ sonic_cli/main.py
