+++ sonic_sfp/sffbase.py
